Clear the code
