+++ src/containers/CurrenciesContainer.js
@@ -3,17 +3,20 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Currency from '../components/Currency';
 
-const CurrenciesContainer = ({ currency }) => (
-  <Container disableGutters>
-    <Grid container>
-      <Currency data={currency} />
-    </Grid>
-  </Container>
-);
+const CurrenciesContainer = ({ currencies }) => {
+  const dataWithoutFirstCurrency = currencies.filter(currency => currency.rank !== '1');
+  return (
+    <Container disableGutters>
+      <Grid container>
+        <Currency data={dataWithoutFirstCurrency} />
+      </Grid>
+    </Container>
+  );
+};
 
 CurrenciesContainer.propTypes = {
   // eslint-disable-next-line react/forbid-prop-types
-  currency: PropTypes.array.isRequired,
+  currencies: PropTypes.array.isRequired,
 };
 
 export default CurrenciesContainer;
